refactor(F2-03): restore domain logic to core as HandleOrderTerminationUseCase
The previous consolidation incorrectly moved domain logic into spring-application.
spring-application handlers should only manage Spring concerns (@Transactional).

- HandleOrderTerminationUseCase (core): all domain logic — position unlock,
  margin release calculation and MarginReleaseEvent publication
- HandleOrderTerminationHandler (spring-application): thin @Transactional wrapper,
  delegates entirely to HandleOrderTerminationUseCase
- RunnerConfig: registers HandleOrderTerminationUseCase bean
